requirements/local.txt
-r base.txt

# Testing
# ------------------------------------------------------------------------------
pytest==7.1.2
pytest-django==4.5.2
factory-boy==3.2.1
faker==13.12.0
coverage==6.4.1

# Code quality
# ------------------------------------------------------------------------------
flake8 # https://github.com/PyCQA/flake8